fix bug find no result for some query

+++ src/core/index/InvertedIndex.h
@@ -85,7 +85,7 @@ struct InvertedListElement {
 
 struct InvertedListIdAndScore {
     unsigned recordId;
-    unsigned score;
+    float score;
 };
